--- app/actions/index.js
@@ -1,6 +1,0 @@
-'use strict';
-import * as TodoActions from './todoActions';
-
-export default {
-  TodoActions
-};
